adds deletion adapter

+++ peachjam/adapters/judgments.py
@@ -1,4 +1,5 @@
 import logging
+import math
 from mimetypes import guess_extension
 from tempfile import NamedTemporaryFile
 
@@ -28,8 +29,7 @@ from peachjam.plugins import plugins
 log = logging.getLogger(__name__)
 
 
-@plugins.register("ingestor-adapter")
-class JudgmentAdapter(RequestsAdapter):
+class BaseJudgmentAdapter(RequestsAdapter):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.client.headers.update(
@@ -47,6 +47,15 @@ class JudgmentAdapter(RequestsAdapter):
         for code in self.court_codes:
             self.filters["court__code"].append(code)
 
+    def client_request(self, method, url, **kwargs):
+        log.info(f"{method.upper()} {url}")
+        r = getattr(self.client, method)(url, **kwargs)
+        r.raise_for_status()
+        return r
+
+
+@plugins.register("ingestor-adapter")
+class JudgmentAdapter(BaseJudgmentAdapter):
     def check_for_updates(self, last_refreshed):
         updated = self.check_for_updated(last_refreshed)
         return updated, []
@@ -236,6 +245,65 @@ class JudgmentAdapter(RequestsAdapter):
                     expression_frbr_uri=expression_frbr_uri
                 ).first()
                 if document:
+                    document.delete()
+            else:
+                raise e
+
+
+@plugins.register("ingestor-adapter")
+class JudgmentDeleteAdapter(BaseJudgmentAdapter):
+    def check_for_updates(self, last_refreshed):
+        from peachjam.tasks import get_deleted_documents
+
+        # kick of bg tasks in batches
+        judgments_count = Judgment.objects.count()
+        log.info(f"Judgments count: {judgments_count}")
+        batch_range = self.settings.get("batch_range") or 10000
+        batch_range = int(batch_range)
+        batches = math.ceil(judgments_count / batch_range)
+        log.info(f"Batches count: {batches}")
+        for r in range(0, batches):
+            get_deleted_documents(
+                self.ingestor.pk, r * batch_range, (r + 1) * batch_range
+            )
+        return [], []
+
+    def get_deleted_documents(self, range_start, range_end):
+        from peachjam.tasks import delete_document
+
+        log.info(f"checking range {range_start} - {range_end}")
+        expression_frbr_uris = (
+            Judgment.objects.filter(court__code__in=self.court_codes)
+            .order_by("pk")
+            .values_list("expression_frbr_uri", flat=True)[range_start:range_end]
+        )
+        url = f"{self.api_url}/validate-expression-frbr-uris"
+        result = self.client_request(
+            "post", url, json={"expression_frbr_uris": list(expression_frbr_uris)}
+        ).json()
+        invalid_frbr_uris = result["invalid_expression_frbr_uris"]
+        log.info(f"found {len(invalid_frbr_uris)} documents to delete")
+
+        qs = Judgment.objects.filter(court__code__in=self.court_codes).values_list(
+            "expression_frbr_uri", flat=True
+        )
+        # find invalid_frbr_uris that exist in the queryset
+        existing = set(invalid_frbr_uris) & set(qs)
+
+        for frbr_uri in list(existing):
+            delete_document(self.ingestor.id, frbr_uri)
+
+    def delete_document(self, expression_frbr_uri):
+        url = f"{self.api_url}/judgment{expression_frbr_uri}"
+        try:
+            document = self.client_request("get", url)
+        except requests.exceptions.RequestException as e:
+            if e.response.status_code == 404:
+                document = Judgment.objects.filter(
+                    expression_frbr_uri=expression_frbr_uri
+                ).first()
+                if document:
+                    log.info(f"deleting document {expression_frbr_uri}")
                     document.delete()
             else:
                 raise e
